requirements.txt
# 核心依赖
PyYAML==6.0.1
psutil==5.9.5
requests==2.31.0
watchdog==3.0.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0

# Web 框架
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2

# 数据可视化
plotly==5.18.0

# 任务队列
redis==5.0.1
celery==5.3.4

# 开发工具
black==23.11.0
isort==5.12.0
mypy==1.7.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1 